cbuildbot: switch to python3 only via virtualenv wrapper [reland]
This doesn't enforce it on imports of scripts/cbuildbot.py yet
because we still have some python2-compat cbuildbot unittests
importing it.  Once this sticks, we can start converting the
cbuildbot modules over too.

While the code hasn't changed since the previous attempt to use
the virtualenv wrapper, we've added the missing mysql wheel that
caused the previous revert.

This should be safe to depend on as we use cbuildbot_launch
to bootstrap the checkout before it runs cbuildbot.

BUG=chromium:997354
TEST=CQ passes
TEST=cbuildbot w/sqlalchemy import & mysql db setup works

Cq-Depend: chromium:2123737
Change-Id: If61d6489dafaf7c09fe8ae9f18196629ede3f138
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2123738
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>