cbuildbot: switch to python3 only via virtualenv wrapper [reland]
This doesn't enforce it on imports of scripts/cbuildbot.py yet
because we still have some python2-compat cbuildbot unittests
importing it.  Once this sticks, we can start converting the
cbuildbot modules over too.

While the code hasn't changed since the previous attempt to use
the virtualenv wrapper, we've added the missing mysql wheel that
caused the previous revert.  We've also improved the virtualenv
creation logic to work with depot_tools non-system python3.

This should be safe to depend on as we use cbuildbot_launch
to bootstrap the checkout before it runs cbuildbot, and we've
fixed the nested execution of virtualenv wrapped programs.

BUG=chromium:997354
TEST=CQ passes
TEST=cbuildbot w/sqlalchemy import & mysql db setup works

Change-Id: Ibe4d76c189112aad55d32837f6624b3b1ac824f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2136320
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>