cbuildbot: switch to python3 only via virtualenv wrapper
This doesn't enforce it on imports of scripts/cbuildbot.py yet
because we still have some python2-compat cbuildbot unittests
importing it.  Once this sticks, we can start converting the
cbuildbot modules over too.

This isn't exactly a reland of the previous attempts as we're
also switching to our virtualenv wrapper.  This allows us to
pull in our own sqlalchemy wheel and not worry about whatever
the host system has (or does not have).

This should be safe to depend on as we use cbuildbot_launch
to bootstrap the checkout before it runs cbuildbot.

BUG=chromium:997354
TEST=CQ passes

Change-Id: I3817d1c1cc70a029d6ace6255e7013c8ba2029cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2120254
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>