trivial: Clean up Tile constructor parameters
Pass pointers to FrameScratchBuffer and DecoderState into the
Tile class to avoid passing several specific parameters from
those two classes. This greatly reduces the number of parameters
to the Tile class.

As a side effect, DecoderState has to be moved to its own file to
avoid a circular dependency (which is another desirable
refactoring in itself).

PiperOrigin-RevId: 299422536
Change-Id: I3600050d129c3f64a016a1fa30fb244eee003117