Add a dynamic buffer helper class
There are multiple places where we store a dynamically sized buffer
along with a size parameter and resize them across frames if the size
increases. Unify all this into a helper class.

There are two versions of the helper class one for unaligned
memory and another one for aligned memory.

PiperOrigin-RevId: 299447741
Change-Id: I88b1cb5cd00698abd7a972455cde455be962fba0