Eliminate the lag induced by intra prediction
When applying the post filters in the decode loop, we lag by one
superblock row to account for intra prediction. That lag can be
eliminated by making a copy of the last row of the current
superblock row to be used for the intraprediction of the next
superblock row.

When computing the top_row for intra prediction, if the top row
is coming from the current superblock row, then we use the
buffer_ as usual. If the top row is coming from the previous
superblock row then we use the copied buffer.

We do this for the single threaded path as well (we do some
extra row copies but it improves cache locality). There is no
significant performance change in that case.

PiperOrigin-RevId: 300224668
Change-Id: Ic36b7933a335ee4fabeedfa49c9a9e622aa55255