gav1_decode: Clear the input_buffer if limit is reached
Once the --limit is reached, we should not enqueue any more
buffers for decoding. If we are holding on to any pending input
buffers that have been read from the file but not yet enqueued
for decoding, clear that when the limit is reached.

Fixes a subtle bug in some cases where we decode an extra frame
(but won't affect the output as we will not write it into the
file).

PiperOrigin-RevId: 300672565
Change-Id: I663e9464c836c0795b2c7a00fa2bfcdf39d30a17