post_filter: Refactor cdef row copying
Refactor the CopyRows function to move the main loop out of it.
The new code is functionally identical to the old one. It just
makes the stride increments read a bit more natural (instead of
magic increments and decrements inside the CopyRows function
based on some booleans).

This is similar to how we populate the loop restoration block.

PiperOrigin-RevId: 300807593
Change-Id: Id44c56d22c28e3c32f62de5e8917e5cecbf10fbd