Fix build of v8 in shared configuration
We don't apply the compiler flag /Zc:dllexportinlines- so that inline
functions are not imported and exported across dll boundaries.  This
flag is normally added in the v8_component template of v8/gni/v8.gni.
The downstream change conditionally undoes this addition of the flag
to prevent the optimization.  The condition is based on the value of
the GN variable remove_dll_export_lines_, which is defined in the v8
v8_component of v8/BUILD.gn.

The removal of this flag has a side effect of requiring exportability
of the data member types.  One of these types is std::shared_ptr, which
depends on __shared_count that is not exportable.  This causes a
problem with the classes v8::CompiledWasmModule and
v8::WasmModuleObjectBuilderStreaming. To workaround this issue, we
don't export the entire class and selectively export the member
functions that are needed to be exported.