Discovery: More public classes
This CL introduces DnsSdServicePublisher and DnsSdServiceFactory into
discovery/public

Previously, @mfoltz/@btolsch/I have discussed this, and decided it was
better to let embedders implement this themselves. The reasoning at the
time was that this class is simple enough that there is little to be
gained from storing this code in a centralized location. Since then,
the following has changed:
- Callback DnsSdPublisher::Client has been added. Most embedders do not
  care about this, so expecting them to implement it will only lead to
  confusion.
- The recent CastAgent CL shows that use of the Discovery pipeline can
  get confusing for those not already familiar with it:
  https://chromium-review.googlesource.com/c/openscreen/+/2102902
- Use of DEPS rules mean that external uses of the pipeline can be
  limited to just discovery/public to encourage embedders to avoid the
  complex internals of the library.

Change-Id: I3fecdebb2e352c4dbf9aabe0489c17f0c8481572
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2103172
Reviewed-by: Max Yakimakha <yakimakha@chromium.org>
Commit-Queue: Ryan Keane <rwkeane@google.com>