Make the intra_prediction_buffer per tile
The intra_prediction_buffer is used to store a copy of the last
row of each superblock row to facilitate the intra prediction of
the next superblock row.

In the current implementation, there is one row of buffer per
frame and each tile uses it's appropriate columns. Change this
behavior so that each tile has its own copy of the buffer.

This change is needed to support tile threading in conjunction
with frame threading. In that case, there might be multiple tile
threads that are running at a given time and they will not be able
to use this shared intra_prediction_buffer.

PiperOrigin-RevId: 301747266
Change-Id: I8496f62bf502d53d0fdaf971957400e143585c97