gerrit: fix some actions after rewrite
The assign & message actions were inheriting from the wrong parent.
Add an assert to catch that in the future (rather than silently
exiting 0), and then fix the inheritance.

Also fix the order of arguments on these to match previous behavior
and to match all the other actions that take an arbitrary number of
CLs.

BUG=chromium:1063185
TEST=`gerrit message -n 1234 foo` works again

Change-Id: Idffac364dd643fe79a95f1b1028d4a66608217ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2111893
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>