Fix crash with FieldTrials
Normally in Chromium, the file descriptor for the field trial shared
memory region is inherited by the child process. Since blpwtk2 embedders
launch the child processes without involvement of Content, the file
descriptors need to be duplicated onto the child process manually.

When blpwtk2 is requested to create a host channel, it passes through
a handle of the child process over to FieldTrialList. FieldTrialList
generates a duplicate file descriptor with respect to the child process
and sends that descriptor instead.