Always use the same site instance
When a user navigates from one domain to another domain, Content may
switch to a different site instance so that a different renderer is use
to load the new page from the new domain. This breaks renderer affinity
requirements of blpwtk2. For this reason, we always use the same site
instance for the same web view.