Fix blpwtk2 URL loader
This commit makes sure that we call NotifyHeadersComplete
asynchronously.  Prior to this change, it was always called
synchronously for URLs that we handled. This was acceptable for
Chromium 76 but Chromium 80 requires the callback to be invoked
asynchronously.

This commit also makes sure that we callout to the resource loader
bridge to load the content body.