Use "call" in gclient.bat to restore title
The title of command prompt windows indicates what command is running.
However it is easy to confuse Windows so that it doesn't clear the
status when a program termintes. Apparently a tail from one batch file
to another is sufficient to do this. This change uses "call" to run
python to avoid this minor glitch, making is so that I can tell when
gclient has completed merely by glancing at the command prompt status
bar.

Change-Id: Ibd310362b8dd78589abc443be1841f0937c791fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/2133303
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>