rtc-svc: Fix to SVC for HD with spatial layers > 1
The seq_params->sb_size is set on key frame/first frame,
and so when #spatial_layers > 1 it should be set based on
the top/highest resolution (oxcf.width/height), instead of
the resolution for the given/base layer (cm->width/height).

This fixed crash for SVC 720p input with #spatial_layers > 1.

Add unittest which catches the issue.

Change-Id: Ia16fabad33aca997b88b7e372063eabcddd5e555