Discovery: Multiple Network Interfaces Part 1
This CL is the first part of supporting multiple network interfaces.
It makes necessary change for duplicate instances of DNS-SD to exist,
each responsible for a single Network Interface.
Additionally, as the endpoint is no longer published as part of the
DnsSdInstanceRecord, these parameters are removed and use of this class
is split into 2 parts - one containing the endpoints and one not.

Future CLs will be responsible for:
- Updating E2E tests.
- Adding support for initiating multiple instances of
  discovery/dnssd/impl/service_impl.h (and renaming it).

For full context, see CL:
https://chromium-review.googlesource.com/c/openscreen/+/2076164

Bug: openscreen:105

Change-Id: I52168f03513d1a605ae3feda50a221f9f505cd79
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2093012
Commit-Queue: Ryan Keane <rwkeane@google.com>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Max Yakimakha <yakimakha@chromium.org>