Use extended-length file path
By default, Windows has a file path limit of 260 characters.  According to
https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file#maximum-path-length-limitation
the limit can be extended to 32,767 by using the extended-length file
path notation.

The toolchain directory contains many files that are deeply nested in
subdirectories. The fullpath for some of the files exceed the 260
character limit. The change in this commit allows for the files to be
accessed correctly.