Reland: device: Don't assume '--device=localhost' means VM.
Some folks have been setting up SSH tunnels locally to route to
a test device. In that case, `cros_run_test` assumes they want to
run the test on a VM and tries launching one.

This fixes that by only assuming that we want to run on a VM if
'--device' is absent all together.

RELAND fix:
 - Adds '--ssh-port' back as an arg when launching a VM.

BUG=chromium:1067800, chromium:1070032
TEST=unittests
TEST=cros_vm --start --ssh-port 9221
TEST=cros_run_test --device localhost:9222
TEST=cros_run_test --ssh-port 9222
TEST=ran this on chrome's bots: crrev.com/c/1863361/40

Change-Id: I1ab4527e8e138f1572df069ad8f90f44d2e37aed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2148147
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>