device: Don't assume '--device=localhost' means VM.
Some folks have been setting up SSH tunnels locally to route to
a test device. In that case, `cros_run_test` assumes they want to
run the test on a VM and tries launching one.

This fixes that by only assuming that we want to run on a VM if
'--device' is absent all together.

BUG=chromium:1067800
TEST=unittests
TEST=cros_run_test --device localhost --port 9222
TEST=ran this on chrome's bots: crrev.com/c/1863361/36

Change-Id: Ic613e7929320b537bffcdca3cea0e4af426d602d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2138195
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>