Add legacy (Debian Jessie) sysroot
Currently, the arm64 build bot fails because the sysroot image
references a version of the standard library that is much too new for
the build bot to run against.

This patch lays the groundwork to resolve that issue by adding a new
platform argument to openscreen, with Jessie sysroots that should fix
the build issue.

Change-Id: Ie2cd1de1a59281eab2d86b78d6d36897722185be
TBR=btolsch@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2140478
Reviewed-by: Jordan Bayles <jophba@chromium.org>