Move from jessie to strech
This patch removes the jessie sysroots, which are too old for the C++
features that we use, in favor of new stretch sysroots that I generated
by pulling an old version of chromium/src and using the
sysroot-creator.sh script with a stretch package configuration.

TBR due to buildbucket.cfg changes, trying to get the bot landed today.

TBR=btolsch@chromium.org

Change-Id: I6dfae987c4be6aae24386ede3abe33d01596821e
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2142317
Reviewed-by: Jordan Bayles <jophba@chromium.org>