Add a new AOMSIGN macro
This commit adds a new macro AOMSIGN, which can be used to get the
sign of an integer. The commit changes to use the macro to replace
"shifting signed 32 bit integer by 31", which is now an undefined
behavior in latest C/C++ standards. The replacement also makes code
easier to understand.

Change-Id: I335135c68fc6583d4e4ab2efa7454a12a782d02b