dlcservice: Add State to outgoing signal.
The states of the dlcservice should be something that is permanent and
queryable. Currently, FAILED is not really queryable because it can
change right away and users miss. This CL adds two states IDLE and
INSTALLING. The users of dlcservice need to figure out the state of
their requested DLC by querying it through GetDlcState().

The current implementation is a middle ground one until we change the
clients to use the new states.

BUG=chromium:1071260
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: I9970dc404336a4cadb506c93610ece2305385c55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2163358
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 026958f29fc02592ca2caba90d6f3eaed262bf8c