perf_uploader: fix various bugs caught by python3
We were using strings when we needed bytes when posting data.

The error uncovered bad exception handling with the wrong name, and
our retry handler assumed it was always a specific exception type.

We switch perf_dashboard_tool to python3.  This is only run by devs,
not by bots, so should be fine.

BUG=chromium:1072008
TEST=`./lib/perf_dashboard_tool -u bytes -t disk.size --cros-version 6689.0.0 --chrome-version 41.6689.0.0 2000400100 --url http://google.com`

Change-Id: I7a6f757ae7f9a64e03131435a51ffe22b05eef02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2154054
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Tiancong Wang <tcwang@google.com>