Revert ARM64 bot to 16.04 due to linker problems
Currently, cross compiling on bionic is borked since there is an
"unsupported GNU_PROPERTY_TYPE" on several of the compiler artifacts.
This warning cannot be turned off or ignored with a flag, so the only
way is to use a different linker version.

This chain will have a follow up patch to force the swarming bot, since
the swarming bot tests MUST be run on 18.04 instead of 16.04.

It's a delicate flow, if 18.04 updates to a more recent linker version
or if we can figure out how to roll our own linker tools in cross
compile this will hopefully go away.

Bug: openscreen/18
Change-Id: I1ccd4eba9d97020c56b0aa9af131dc673bd6b466
TBR=btolsch@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2158819
Reviewed-by: Jordan Bayles <jophba@chromium.org>