afdo: Migrate CreateAndUploadMergedAFDOProfile to recipe
The legacy benchmark-afdo-generate is still using a cbuildbot
function CreateAndUploadMergedAFDOProfile() in afdo.py that merges
recent AFDO profiles to provide stability to Android/Linux. This
patch rewrites this function into a recipe world, reusing the
functions, as well as preserving all the nice tests from the old
world.

The major change is this function itself will not compress and
upload the merged profile. Instead, leave the part to the caller.

BUG=chromium:1019868
TEST=Unittest pass and will use legacy builder for testing

Change-Id: Ib4b0291f25c4c28087ea9a46ebfabc4e162881c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2159597
Commit-Queue: Tiancong Wang <tcwang@google.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>