Readd Debian stretch for bot compilation
Currently the ARM64 bots are using too new of a version of glib, so they
are incompatible with all of the bots in the luci.flex.try pool that run
ARM64.

This patch readds the Debian stretch sysroot for cross compiling, so we
link an older version of glib that is either the same or forwards
compatible with the newer version linked in the Ubuntu 18.04 bots.

Bug: openscreen/18
Change-Id: I3ada49ac5197844b0c49028194a17c993c7b2663
TBR=btolsch@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2163436
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>