Replace release with buildmaster configuration (#770)
Prior to this commit, the `static_release` configuration had the
`is_official_build` flag turned off while the `static_buildmaster`
configuration had it turned on.  This flag enables a class of
optimizations that should be turned on for production builds.

The `make_devkit.py` script uses the `static_release` configuration but
manually enables the flag by adding it to the `GN_DEFINES` environment
variable. This produces an output directory called `static_release`
while having the characteristics of a build master configuration.

Recently there was a regression introduced by commit `2f5d61f4`. This
change was made to allow for the generation of builds against
dynamically loaded CRT. A side effect of the change caused the
`is_official_build` flag to be explicitly turned off for the
`static_release` configuration. This caused the devkits generated by
the `make_devkit.py` script to produce a release build rather than a
buildmaster build.

To fix this problem, we are removing the `static_buildmaster`
configuration and we are always enabling the `is_official_build` flag
in the `static_release` configuration. This will give us two advantages:

 1. Move the logic of whether all optimizations are enabled into the
    `blpwtk2.py` script. Before the logic was split between this script
    and the `make_devkit.py` script.
 2. Allow us to test a fully optimized build by using the
    `static_release` configuration. Prior to this commit, we needed to
    rebuild using the `static_buildmaster` configuration when we wanted
    a fully optimized build.