Fix builders by bringing back chromeos_config_unittest symlink.
Fix the config-updater and luci-scheduler builders by bringing back the
symlink the builders invoke these with. These were removed in
https://crrev.com/c/2166515 and the correct solution is likely invoking
this via pytest - for now to quickly get them going again just bring
back this symlink.

For the failures that this is causing in production see:

https://ci.chromium.org/p/chromeos/builders/general/Infra/b8881821635947082176?

BUG=b:155162381
TEST=cros_sdk -- /mnt/host/source/chromite/config/chromeos_config_unittest
TEST=cros_sdk -- /mnt/host/source/chromite/config/chromeos_config_unittest --update
TESt=run_tests

Change-Id: I3adcb487b2ac3214a5de57257d9c765ffa31de73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2170302
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>
Tested-by: David Burger <dburger@chromium.org>