cros_chrome_sdk: Explicitly set use_goma_thin_lto to false
The gn argument use_goma_thin_lto causes ThinLTO code generation
to be distributed on Goma. The gn configuration checks that when
this is enabled, use_thin_lto and use_goma are also true, and
asserts if not. Since use_thin_lto can have different defaults in
different parts of the build, this can cause the assert to trigger.
Since use_goma_thin_lto is not intended to be used with simplechrome
at this stage, just explicitly disable it for now.

BUG=chromium:1065073
TEST=cros chrome-sdk --board=nocturne

Change-Id: I9e8bbc0cf6bb7b7e3a218cc325f49d90f6f3f631
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2171161
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>