cbuildbot: remove health_alert properties from configs.
The credentials for sending emails through cbuildbot alerts have been
removed and attempting to send health alerts causes an error in the
report stage and for alerts to not be sent.

In a following change repurpose the code for sending health alerts to
use luci-notify.

BUG=chromium:900740
TEST=./config/refresh_generated_files
TEST=./run_tests

Cq-Depend: chromium:2175207
Change-Id: I00e6f99ecd1daa1b5d141da5b2483355fcc2ef84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2175205
Tested-by: Navil Perez <navil@google.com>
Commit-Queue: Navil Perez <navil@google.com>
Reviewed-by: George Engelbrecht <engeg@google.com>