Disable bvt-arc suite for VMRVC PFQ runs
Currently we are not running any tests on the bvt-arc test suite on RVC
and the suite fails because there are no tests in it ('enumeration
error'). With this CL we temporarily disable the bvt-arc suite by
specifically filtering it out from the returned list of hwtest suites
only on VMRVC boards.

BUG=b:153392483
TEST=`./config/refresh_generated_files`

Change-Id: I2786fdb48d4a4eab720ad134212bec6f12110426
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2172576
Tested-by: Morg <morg@chromium.org>
Auto-Submit: Morg <morg@chromium.org>
Reviewed-by: Lepton Wu <lepton@chromium.org>
Reviewed-by: David Burger <dburger@chromium.org>
Commit-Queue: David Burger <dburger@chromium.org>