pylintrc: Ignore pytest import
Pytest is always available in the virtualenv when we run tests, but the
presubmit linter doesn't run from that virtualenv so it can't import
pytest.

Ignore attempts to import pytest so we don't have to decorate every
usage of it with an ignore directive.

BUG=None
TEST=bare `import pytest` no longer triggers a lint

Change-Id: I941a11d39f027f8fff05eb330e491f691660dd1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2173305
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>