cros_install_debug_symbols: Fix --list call.
The script relies on calling itself with the --list argument as a base
case to avoid instantiating an object that cannot be serialized before
executing the parallel install step. When a package is given, the
appended --list option is parsed as another package. Instead, inject
--list as the first argument.

BUG=chromium:1081458
TEST=manual

Change-Id: Iee47c195a6993dca2d93db3fb0a291d0f1a29943
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2195181
Tested-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>