cbuildbot: Create annotation class for email_notify property
The current buildbot annotations cast all arguments into strings and
replace '@' with 'AT'. This format is not recognized by luci-notify and
alerts are not able to fire.

In order to prevent breaking other properties which may require the '@'
to 'AT' conversion, this change creates a new annotation class
specifically for setting the email_notify property in a format
recognized by luci-notify.

This change allows email_notify to maintain its type and keeps the
email addresses intact.

BUG=chromium:1067014
TEST=./run_tests

Change-Id: I00c27a6105d3075695f16f2c45115ed454a31741
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2197195
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>
Tested-by: Navil Perez <navil@google.com>
Commit-Queue: Navil Perez <navil@google.com>