Fix boundary conditions on string operations (#777)
A previous commit introduced a regression where a closing brace was
inserted into a string buffer one character too early. This effectively
made the output string a malformed JSON string.

The tracing controller in v8 platform indirectly calls the function
ConvertableToTraceFormatWrapper::AppendAsTraceFormat to wrap each trace
argument with curly braces and this issue caused the generated output to
have its last character replaced by the closing brace.

The CPU profiler uses the v8 tracing controller to trace the event
"ProfileChunk" under the category v8.cpu_profiler. The v8 profile
generator collects the JS callstack from the tick sample passed by the
main thread and embeds the callstack into the ProfileChunk event prior
to tracing it.

This ProfileChunk event is sent by the devtools agent to the devtools
frontend when the performance profiler is started. The devtools frontend
uses the embedded callstack from the ProfileChunk event to display call
tree in the performance tab.

The malformed JSON string caused the CPU profiler to not include the
callstack generated by the profile generator into the ProfileChunk
event. This caused the call tree to not appear in the devtools
frontend.

Test plan:
1. Run content_shell and load any web page that continuously runs some
   JS code
2. Bring up the inspector and navigate into the performance tab
3. Start recording for a few seconds and stop.
   The "Main thread" in the graph should contain a call tree of the JS
   function.