Disable "FontSrcLocalMatching" feature (#779)
This feature was recently enabled by default. It's not compatible how the Bloomberg custom fonts are handled, and causes a problem in Windows 7 and 8.1 that fails to fallback to the right fonts specified in the font configuration css by `@font-face` rule, resulting in rendering everything in Times New Roman. 

This commit also reverted the previous fix made for Windows 10 since it caused the fixed font to render wrong when bold. It's not necessary with the feature disabled.