Turn off mDNS support (#782)
Ideally an mDNS server is a system-wide service that listens for queries
from other nodes on the subnet and responds back with a list of services
provided by the system. Enabling the mDNS support in Content seems to
bring up a mDNS server on port 5353. This may be acceptable in ChromeOS
but a zeroconf server like this should be external to the browser in a
typical OS. This commit turns off mDNS support. This disablement was
necessary to prevent a warning from Windows Defender that port 5353 was
being opened.