Fix text distoration when y/x aspect ratio is > 1 (#790)
LayerImpl::GetIdealContentsScaleAndAspectRatio now returns a std::pair
of [x-scale, scale aspect ratio]. An upstream change
(https://chromium-review.googlesource.com/c/chromium/src/+/1790958)
introduced a cap to the scale factor so that it tries to return the
larger of the two scale factors (of the two axis) but never exceeds 5
times the minimum of the two scale factors. In either case, the logic
considers both scale factors when PictureLayerImpl expects the first
of the returned pair to be based only on the x-scale factor.

Suppose the desired scale factor is 1x5. If the capping logic was used,
the returned pair would be (5,5) when it should be (1,5). This will
cause PictureLayerImpl to look for a tile of size (5,25) instead of
(1,5). This triggers the post scaler to shrink the rasterized image
down to the correct size and this double transformation causes a
distoration.

This change-set also fixes non-uniform scaling when GPU process side
rasterization is enabled. This mode can be enabled with
--enable-oop-rasterization but it's currently disabled by default.