Fixed rubberband logic to handle scroll properly (#794) (#798)
There was an upstream change - https://chromium-review.googlesource.com/c/chromium/src/+/1683951 that changed the meaning of the `PaintLayer::Location` and renamed to `LocationWithoutPositionOffset`. After this change, the caller is supposed to check if there is a containing layer and subtract the offset if so. In this PR, the rubberband logic is adjusted to reflect the change.