Merge to M84: Prevent pointer from being sent in the clear over SCTP.
We were using the address of the SctpTransport object as
the sconn_addr field in usrsctp, which is used to get access to
the SctpTransport object in various callbacks.

However, this address is sent in the clear in the SCTP cookie,
which is undesirable.

This change uses a monotonically increasing id instead, which
is mapped back to a SctpTransport using a SctpTransportMap helper
class.

TBR=hta@webrtc.org

Bug: chromium:1076703
Change-Id: I5c6a44801293e3b0aacd032f16f41802f4fecf6d
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/176422
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Tommi <tommi@webrtc.org>
Commit-Queue: Taylor <deadbeef@webrtc.org>
Cr-Original-Commit-Position: refs/heads/master@{#31449}
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/177330
Reviewed-by: Taylor <deadbeef@webrtc.org>
Cr-Commit-Position: refs/branch-heads/4147@{#9}
Cr-Branched-From: 2b7d96959916306cb267899a219cbd20ec24b841-refs/heads/master@{#31262}