embedder-ipc: Install a gin::MultiHeapTracer for 'browser V8' mode (#813)
When initializing 'browser V8' mode, create a `gin::MultiHeapTracer`.
A complementary change to the 'multiple-heap-tracer' branch modifies
the `addV8HeapTracer()`, `removeV8HeapTracer()`, and `setIsolate()` functions
to use this heap tracer when it is is available.

Note that this PR adds the `MultiHeapTracer` to the `gin` namespace, even
though it would be more appropriate for that to be done in the
'multiple-heap-tracer' branch itself. However, since this 'embedder-ipc'
branch introduces the `gin` namespace into 'blpwtk2_toolkitimpl.h',
attempting to also introduce it into the other branch caused git merge
conflicts.