blpwtk2: Introduce blpwtk2::WebView::setNCHitTestRegion() (#819)
The purpose of this new API is to allow an embedder to delegate non-client
hit-testing to the embedding window when that window and the webview are
owned by the same thread. Such an ownership arrangement allows a child window
to return `HTTRANSPARENT` from the `WM_NCHITTEST` message to delegate hit-testing
to its parent window without any problems (Win32 will not delegate to the
parent window when they are owned by separate threads, causing an endless,
essentially futile stream of `WM_NCHITTEST` messages to be delivered to the
child window until the user moves the mouse).

The new API, `setNCHitTestRegion()`, accepts a Win32 region (a `HRGN`) that
is used during processing of the `WM_NCHITTEST` message. If the hit-test point
is within this region, then, the implementation of `WM_NCHITTEST` will return
`HTTRANSPARENT`; if the hit-test point is not within this region, then, the
implementation of `WM_NCHITTEST` will return `HTCLIENT` (and ordinary DOM
events will generated instead).