renderer-ui: Prevent early creation of upstream webview compositor (#826)
The Chromium compositor specifies an abstraction called `cc::LayerTreeFrameSink`
to render frames into. The 'renderer UI' mode includes an additional
implementation of this abstraction that `RenderWebView` creates to allow
compositing into an HWND owned by the renderer thread.

`RenderWebView` creates the proper `cc::LayerTreeFrameSink` when
`notifyRoutingId()` is called as part of the response to a mojo message sent by
the browser.

Depending upon when this message is received, it is possible that upstream
Chromium code will have already created a non-'renderer UI'
`cc::LayerTreeFrameSink`; when this happens, nothing will appear in the webview.

To force the 'renderer UI' implementation to be used, `notifyRoutingId()` calls
`ReleaseLayerTreeFrameSink()` which will force the creation of the expected
implementation.

It would be better if Chromium just did not create the non-'renderer UI'
`cc::LayerTreeFrameSink`, in order to avoid the possibility that the process of
destroying and creating the frame sink puts the webview's compositor into an
unexpected state that it cannot recover from.

If the `content::WebView` is created with the 'initially_hidden' property set
to `true`, then, the non-'renderer UI' `cc::LayerTreeFrameSink` will never be
created, because the webview will not enter a 'visible' state until
`notifyRoutingId()` is called. This PR sets this value to be the inverse
of the `initiallyVisible` parameter of `WebViewImpl`'s constructor (and,
in fact, blpwtk2 always passes 'false' for this parameter).

Furthermore, this PR removes the sequence in `notifyRoutingId()` that forces
a refresh of the `cc::LayerTreeFrameSink`.