renderer-ui: Prevent early creation of upstream webview compositor (#834)
The Chromium compositor specifies an abstraction called `cc::LayerTreeFrameSink`
to render frames into. The 'renderer UI' mode includes an additional
implementation of this abstraction that `RenderWebView` creates to allow
compositing into an HWND owned by the renderer thread.

`RenderWebView` creates the proper `cc::LayerTreeFrameSink` when
`notifyRoutingId()` is called as part of the response to a mojo message sent by
the browser.

Depending upon when this message is received, it is possible that upstream
Chromium code will have already created a non-'renderer UI'
`cc::LayerTreeFrameSink`; when this happens, nothing will appear in the webview.

To force the 'renderer UI' implementation to be used, `notifyRoutingId()` calls
`ReleaseLayerTreeFrameSink()` which will force the creation of the expected
implementation.

It would be better if Chromium just did not create the non-'renderer UI'
`cc::LayerTreeFrameSink`, in order to avoid the possibility that the process of
destroying and creating the frame sink puts the webview's compositor into an
unexpected state that it cannot recover from.

If the `content::WebView` is created with the 'initially_hidden' property set
to `true`, then, the non-'renderer UI' `cc::LayerTreeFrameSink` will never be
created, because the webview will not enter a 'visible' state until
`notifyRoutingId()` is called. This PR sets this value to be the inverse
of the `initiallyVisible` parameter of `WebViewImpl`'s constructor (and,
in fact, blpwtk2 always passes 'false' for this parameter).

Furthermore, this PR removes the sequence in `notifyRoutingId()` that forces
a refresh of the `cc::LayerTreeFrameSink`.

### Changes from earlier version

Additionally, this PR includes remedies for issues with [an earlier version
of this change](https://bbgithub.dev.bloomberg.com/uiinf/chromium.wtk2/pull/826):

* Blink maintains visibility of the 'page' that is separate from the
  visibility of the 'webview', and which is used by Chromium to perform
  certain tasks when the user switches tabs, such as throttling timers.
  `blpwtk2::RenderWebView` has been setting 'webview' visibility but has
  not been setting 'page' visibility, and, since this PR makes
  webview's initially visible, the original version of these changes
  always throttled timers, because the 'page' would never enter a
  'visible' state.

  So, this version of the PR sets page visibility when the webview
  visibility is changed, and in an order that is consistent with
  regular Chromium.

* The original version of this PR caused non-hybrid mode webviews
  to 'flash white' for a noticeable period of time as the initial
  content finished loading. To alleviate this, `blpwtk2::WebViewImpl`
  sets the background color of the webview to black as soon as it
  is created (without setting this, the function
  `content::DelegatedFrameHostClientAura::DelegatedFrameHostGetGutterColor()`
  will use white as the color to display before the webview finishes
  loading).