Log JS exception error and avoid null ScriptState crash in case of additional web script context (#884)
in case of additional web script context.

Note that in a web script context, a document may be associated with a local DOM Window without a frame. In this case, it is possible to get a null ScriptState when reporting a JS exception. This change add guarding code to avoid null ScriptState crash.