Remove unnecessary call to CallMsgFilter (#887)
Prior to Chromium 34, the browser process relied on this message hook
for drag & drop functionality. The dependent code has been removed as
part of https://crbug.com/330735 but the call to this unused message
hook function remained stale. This is not only a dead code but induces
a notable performance hit.