Register metrics with the embedder and support sampling of them through blpwtk2::Toolkit::Delegate (#1028)
Provide a means by which `blpwtk2::Toolkit` can reach out to its embedder to register metrics that can then be periodically sampled by the embedder.

New methods `registerMetric` and `unregisterMetrics` are added to the `blpwtk2::ToolkitDelegate`. When the implementation of `blpwtk2::ToolkitDelegate` is supplied by the embedder, `blpwtk2::Toolkit` can call `registerMetric` on the delegate for each metric it wishes to make available to the embedder. The integer that is returned for each is used as a token to identify that metric when the embedder subsequently periodically calls `blpwtk2::getMetrics`.

As a demonstration, the ability to count `blink::ComputedStyle` objects is added, as well as reporting of `WTF::Partitions` memory metrics.