Disable in-process stack traces (#1073)
* Disable in-process stack traces

In-process stack traces causes Content to query symbol resolutions when
a process crashes. This unnecessarily opens up network requests which
can easily be deferred offline. For blpwtk2, the embedder is expected to
facilitate the symbol resolutions offline after a process crash.

For security reasons this feature is disabled for official builds (aka
static_release). This commit just makes sure that it is also disabled
for shared build configuration for consistency.

* address review comments