Allow embedder to set the path of temp folder (#1103)
Currently blpwtk2 creates temp folders in the system temporary folder.
However, the embedder may require the log files to be stored in a different
location.

This commit allows embedder to configure the temporary folder in the toolkit
creation params. This temp path will be saved in the `Statics::tempFolderPath`,
which is then being used in `DevToolsFrontendHostDelegateImpl` and
`BrowserContextImpl`. 

If embedder doesn't set the temporary folder through toolkit creation params,
the `Statics::tempFolderPath` defaults to the system temp folder.