Fix paint artifacts in antialiased text
In some cases, a certain row in a list or table needs its own graphics
layer. To make sure the remaining rows are on top of this row, the
display list items that pertain to these remaining items are played back
on a squashed compositor layer attached to the same
CompositedLayerMapping.

Note that the main graphics layer was created for this one-off row but
the majority of the content sits on top of the attached squashed layer,
which actually extends beyond the bounds of the main graphics layer.
To make sure the squashed layer gets the correct text background, we
need to search for the next parent LayoutObject and use its background
instead.

Prior to this commit, the background of the main graphics layer of
the non-first compositor layer was used as the destination background
when blitting the lcd text. With this commit, the background of the
parent of the main graphics layer of this non-first compositor layer is
used for the squashed layer only.