lcd-text: add support for translucent background
Prior to this changeset, the text painter in skia used the provided
lcd text background only if the destination pixel was completely
transparent. This didn't work well when the destination was partially
transparent.

With this changeset, the current destination pixel is blended on top of
the provided lcd text background and the resultant color is used as the
background to paint the text.