Copy SkBitmap when scheduling an async call
Prior to this commit, blpwtk2::RenderWebView::StartDragging had
const-ref parameters that were passed onto base::Bind(). This is a
problem because references are essentially raw pointers and there is no
guarantee that the raw pointers will stay valid between the time the
task is posted and the task is executed.

This commit simply makes the const-ref parameters into value types. This
should not incur a performance hit because SkBitmap internally stores
the pixel buffer in a shared pointer. Copying a SkBitmap causes the
pixel buffer to be shared and not copied.