Disable sandbox (#297)
In Chromium 92, the support for enabling sandbox with different process
image was removed. This means that it is now a requirement for all
process images to be identical for sandbox to work.

For now the embedders of blpwtk2 do not support using the same process
image for both the child process and the parent process. Once that
support is implemented, we can remove this disablement of sandbox.