renderer-ui: Mask away some DCHECKs (#308)
This is not a good practice but renderer-ui is already breaking so many
assumptions made by upstream code and I don't think it's worth
investigating debug assertions unless it causes a real problem in
production builds. This branch will soon go away so these DCHECKs will
be re-enabled soon as well.