RendererUI: Don't access the filesystem manager
Currently, the embedders of blpwtk2 rely only on the filename topic to
be present in the drag & drop data; It doesn't need the browser to
provide a handle to the actual file data.

In RendererUI mode, the file system access manager is not available.
Since we don't need the browser to provide the renderer with a handle to
the file data, we conditionally don't request a token if the file system
access manager is not available.