Create NativeTheme when blpwtk2 toolkit is created (#362)
Content shell normally creates this singleton instance of NativeThemeWin
when the first request to create a webview comes in. Since `content`
creates the renderer process **after** the webview creation request
comes in, the call to ThemeHelper::SendSystemColorInfo happens to be
after the system colors are initialized.

For blpwtk2, the externally-managed renderers are created and linked to
the browser process before the webview creation request comes in. This
causes ThemeHelper::SendSystemColorInfo to be called much earlier. To
ensure the system colors are initialized before this call, we create the
singleton instance right here, which is before any renderer process is
linked with the browser process.

This change-set contains an additional change in blpwtk2_shell. Prior to
this change, the blpwtk2_shell renderer processes wait 0.5 seconds after
they are launched to get the channel id from the browser process. If it
doesn't get a channel id, the renderer process falls back to use an
in-process browser thread instead of connecting to the external browser
process.

Normally 0.5 seconds is plenty of time for the browser to provide the
channel id via shared memory but this becomes a problem when the browser
process is paused for debugging. In order to make sure the renderer
always connects to the expected browser process, it now polls every 0.5
seconds for a channel id until it gets one.